refactor: removing model from graphicalEditor
