+++ HealthcareSystem/GraphicalEditorServer/DTO/DrugDTO.cs
@@ -15,23 +15,13 @@ namespace GraphicalEditorServer.DTO
 
         public DrugDTO() { }
 
-        public DrugDTO(DrugTypeDTO drugTypeDTO, string name, int quantity, DateTime expirationDate, string producer)
+        public DrugDTO( string name, int quantity, DateTime expirationDate, string producer, DrugTypeDTO drugTypeDTO)
         {
             this.DrugTypeDTO = drugTypeDTO;
             this.Name = name;
             this.Quantity = quantity;
             this.ExpirationDate = expirationDate;
             this.Producer = producer;
-        }
-
-
-        public DrugDTO(DrugDTO drug)
-        {
-            this.DrugTypeDTO = drug.DrugTypeDTO;
-            this.Name = drug.Name;
-            this.Quantity = drug.Quantity;
-            this.ExpirationDate = drug.ExpirationDate;
-            this.Producer = drug.Producer;
-        }
+        }        
     }
 }
