+++ HealthcareSystem/GraphicalEditorServer/DTO/DrugTypeDTO.cs
@@ -12,19 +12,13 @@ namespace GraphicalEditorServer.DTO
         public string Purpose { get; set; }
 
         public DrugTypeDTO() { }
-
-        public DrugTypeDTO(string type, string purpose)
-        {
-
+       
+        public DrugTypeDTO(int id, string type, string purpose) {
+            this.Id = id;
             this.Type = type;
             this.Purpose = purpose;
         }
 
-        public DrugTypeDTO(DrugTypeDTO drugType)
-        {
-
-            this.Type = drugType.Type;
-            this.Purpose = drugType.Purpose;
-        }
+       
     }
 }
