+++ HealthcareSystemFrontends/GraphicalEditor/Service/DrugService.cs
@@ -2,7 +2,6 @@
 
 using GraphicalEditor.DTO;
 using GraphicalEditor.Models;
-using GraphicalEditor.Models.Drugs;
 using RestSharp;
 using System;
 using System.Collections.Generic;
