+++ HealthcareSystemFrontends/GraphicalEditor/Service/DrugTypeService.cs
@@ -1,4 +1,4 @@
-﻿using GraphicalEditor.Models.Drugs;
+﻿using GraphicalEditor.DTO;
 using RestSharp;
 using System;
 using System.Collections.Generic;
