+++ HealthcareSystemFrontends/GraphicalEditor/Service/PatientService.cs
@@ -1,6 +1,5 @@
 ﻿using GraphicalEditor.DTO;
 using GraphicalEditor.Models;
-using GraphicalEditor.Models.Drugs;
 using RestSharp;
 using System;
 using System.Collections.Generic;
